Use databaseConfig.
